refactor: cqrs 개념 적용해 조회쿼리와 커맨드 쿼리 분리
- Notice JoinColumn,mappedBy 추가
- BaseTimeEntity, BaseEntity 외부에서 호출 못하도록 abstract로 변경
