+++ src/main/java/com/hongikbros/jobmanager/common/domain/BaseTimeEntity.java
@@ -12,7 +12,7 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener.class)
-public class BaseTimeEntity implements Serializable {
+public abstract class BaseTimeEntity implements Serializable {
 
     @CreatedDate
     private LocalDateTime createdDate;
