+++ src/main/java/com/hongikbros/jobmanager/notice/command/application/NoticeService.java
@@ -1,16 +1,16 @@
-package com.hongikbros.jobmanager.notice.application;
+package com.hongikbros.jobmanager.notice.command.application;
 
-import static com.hongikbros.jobmanager.notice.application.exception.NoticeServiceExceptionCode.*;
+import static com.hongikbros.jobmanager.notice.command.exception.NoticeServiceExceptionCode.*;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.hongikbros.jobmanager.notice.application.dto.NoticeResponse;
-import com.hongikbros.jobmanager.notice.application.exception.NotAllowedEmptyUser;
-import com.hongikbros.jobmanager.notice.domain.NoticeRepository;
-import com.hongikbros.jobmanager.notice.domain.notice.Notice;
-import com.hongikbros.jobmanager.notice.domain.scraper.Scraper;
-import com.hongikbros.jobmanager.notice.ui.dto.NoticeCreateRequest;
+import com.hongikbros.jobmanager.notice.command.dto.NoticeResponse;
+import com.hongikbros.jobmanager.notice.command.exception.NotAllowedEmptyUser;
+import com.hongikbros.jobmanager.notice.command.domain.NoticeRepository;
+import com.hongikbros.jobmanager.notice.command.domain.notice.Notice;
+import com.hongikbros.jobmanager.notice.command.domain.scraper.Scraper;
+import com.hongikbros.jobmanager.notice.command.dto.NoticeCreateRequest;
 import com.hongikbros.jobmanager.security.core.CurrentMember;
 
 @Service
