+++ src/main/java/com/hongikbros/jobmanager/notice/command/domain/NoticeRepository.java
@@ -1,9 +1,8 @@
-package com.hongikbros.jobmanager.notice.domain;
+package com.hongikbros.jobmanager.notice.command.domain;
 
+import com.hongikbros.jobmanager.notice.command.domain.notice.Notice;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import com.hongikbros.jobmanager.notice.domain.notice.Notice;
 
 @Repository
 public interface NoticeRepository extends JpaRepository<Notice, Long> {
