+++ src/main/java/com/hongikbros/jobmanager/notice/ui/NoticeController.java
@@ -2,17 +2,16 @@ package com.hongikbros.jobmanager.notice.ui;
 
 import java.net.URI;
 
+import com.hongikbros.jobmanager.notice.query.applicaion.NoticeViewService;
+import com.hongikbros.jobmanager.notice.query.dto.NoticeResponses;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import com.hongikbros.jobmanager.common.core.validation.ValidationSequence;
-import com.hongikbros.jobmanager.notice.application.NoticeService;
-import com.hongikbros.jobmanager.notice.application.dto.NoticeResponse;
-import com.hongikbros.jobmanager.notice.ui.dto.NoticeCreateRequest;
+import com.hongikbros.jobmanager.notice.command.application.NoticeService;
+import com.hongikbros.jobmanager.notice.command.dto.NoticeResponse;
+import com.hongikbros.jobmanager.notice.command.dto.NoticeCreateRequest;
 import com.hongikbros.jobmanager.security.core.AuthMember;
 import com.hongikbros.jobmanager.security.core.CurrentMember;
 
@@ -23,9 +22,11 @@ public class NoticeController {
     public static final String API_NOTICE = "/api/notice";
 
     private final NoticeService noticeService;
+    private final NoticeViewService noticeViewService;
 
-    public NoticeController(NoticeService noticeViewService) {
-        this.noticeService = noticeViewService;
+    public NoticeController(NoticeService noticeService, NoticeViewService noticeViewService) {
+        this.noticeService = noticeService;
+        this.noticeViewService = noticeViewService;
     }
 
     @PostMapping
@@ -38,5 +39,12 @@ public class NoticeController {
         return ResponseEntity
                 .created(URI.create(API_NOTICE + "/" + notice.getId()))
                 .body(notice);
+    }
+
+    @GetMapping()
+    public ResponseEntity<NoticeResponses> findAll(@AuthMember CurrentMember currentMember) {
+        NoticeResponses noticeResponses = noticeViewService.findAllByMemberId(currentMember);
+
+        return ResponseEntity.ok(noticeResponses);
     }
 }
